// web/projects/portal/src/app/portal/data/data-navigation-tree/portal-data-type.ts
/*
 * This file is part of StyleBI.
 * Copyright (C) 2024  InetSoft Technology
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
/**
 * Portal data types.
 */
export enum PortalDataType {
   DATA_SOURCE = "DATA_SOURCE",
   XMLA_SOURCE = "XMLA_SOURCE",
   DATA_SOURCE_FOLDER = "DATA_SOURCE_FOLDER",
   DATA_SOURCE_ROOT_FOLDER = "DATA_SOURCE_ROOT_FOLDER",
   DATABASE = "DATABASE",
   EXTENDED_PARTITION = "EXTENDED_PARTITION",
   EXTENDED_LOGIC_MODEL = "EXTENDED_LOGIC_MODEL",
   FOLDER = "FOLDER",
   LOGIC_MODEL = "LOGIC_MODEL",
   PARTITION = "PARTITION",
   PRIVATE_WORKSHEETS_FOLDER = "PRIVATE_WORKSHEETS_FOLDER",
   SHARED_WORKSHEETS_FOLDER = "SHARED_WORKSHEETS_FOLDER",
   VPM = "VPM",
   VPM_FOLDER = "VPM_FOLDER",
   DATA_MODEL = "DATA_MODEL",
   DATA_MODEL_FOLDER = "DATA_MODEL_FOLDER"
}